pyproject.toml
[tool.poetry]
name = "am-flow-test"
version = "0.1.0"
description = "AM Flow experiment"
authors = ["Domenico Nappo <domenico.nappo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
slumber = "^0.7"
python-dotenv = "^1.0"
flask = "^2.3"

[tool.poetry.dev-dependencies]
pytest = "^7.4"
flake8 = "^6.0"
black = "^23.7"
isort = "^5.12"
mypy = "^1.4"
ipdb = "^0.13"

[tool.poetry.extras]
dev = ["ipython"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
python_files = "tests.py test_*.py *_tests.py"
pythonpath = ["src", "tests"]  # adds /src and /tests directories to the pythonpath

[tool.mypy]
mypy_path = ["src", "tests"]
python_version = "3.10"
disallow_untyped_defs = true
show_error_codes = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
plugins = ["mypy_django_plugin.main"]

strict_optional = true

[tool.django-stubs]
django_settings_module = "openai_ms.settings"
